eval fx manhattan generator float64

+++ evalfx.go
@@ -48,10 +48,10 @@ func astar_hamming(pos, startPos, goalPos [][]int, path []byte) int {
 	return len(path) + 1 + greedy_hamming(pos, startPos, goalPos, path)
 }
 
-func astar_manhattan_generator(weight int) evalFx {
+func astar_manhattan_generator(weight float64) evalFx {
 	return func(pos, startPos, goalPos [][]int, path []byte) int {
 		initDist := len(path) + 1
-		return initDist + weight*greedy_manhattan(pos, startPos, goalPos, path)
+		return initDist + int(weight*float64(greedy_manhattan(pos, startPos, goalPos, path)))
 	}
 }
